requirements.txt
streamlit~=1.10.0
altair~=4.1.0
pandas~=1.2.3
tinydb~=4.7.0
streamlit-authenticator~=0.1.5
PyYAML~=6.0